RAG/RAGDEMO/backend/requirements.txt
fastapi
uvicorn
pydantic
chromadb
langchain>=0.1.17
langchain-groq
langchain-core
sentence-transformers
python-bidi
python-arabic-reshaper
fuzzywuzzy
groq
motor
python-dotenv
passlib
passlib[bcrypt]
pymongo
elevenlabs
python-multipart
python-dateutil
beautifulsoup4
playwright
pytz
bson
httpx
requests
beautifulsoup4
playwright
